fixed more warnings, setup main to run on testset

+++ code/java/lire_indexer_searcher/src/main/java/classifier/WeightedImageSearchHitClassifier.java
@@ -64,7 +64,7 @@ public class WeightedImageSearchHitClassifier {
                 }
             }
             for(Prediction pred : predictionsOfClass){
-                pred.score = pred.score * this.weights.getWeight(className,pred.category) / hits.length();
+                pred.score = pred.score * this.weights.getWeight(className,pred.category) / (double)hits.length();
             }
             allPredictions.add(predictionsOfClass);
         }
